--- Sync ----Added the Sync and SyncServer protocol.  If enabled and set upcorrectly, this will attempt to Sync bans across a network of any amountof servers.
--- Requirements ---A Bukkit server may be configured to be the
"SyncServer", or just a"Syncer". Additionally, SyncServer.class may be
run as a standaloneserver (without needing a Bukkit server to host it)
using a commandlike:java org.maxgamer.maxbans.sync.SyncServer -pass
PASSWD -port PORT#

A sync setup requires exactly one server and any amount of clients.

--- Protocol ---The sync server protocol is relatively easy, and uses
sockets. Stringsare converted to bytes using the ISO-8859-1 CharSet
(E.g. "HelloWorld".getBytes("ISO-8859-1");). The bytes are then sent
through thesocket, and received, and converted back to String form using
theISO-8859-1 CharSet again.  The CharSet is important, otherwise
specialcharacters can break!

A typical conversation goes something like this:(Client makes connection
request to server through a regular socket)[Client => Server] @connect
-pass MY_ENCRYPTED_PASSWORD //Sends passwd[Server => Client] @connect
//Receiving @connect from the server means success![Client => Server]
@ban -name NewGuy -reason Potty mouth -banner MrGrumpyGills
-broadcast[Server => OTHER_CLIENTS] @ban -name NewGuy -reason Potty
mouth -banner MrGrumpyGills[Client => Server] @disconnect -reason Plugin
is being disabled(Client drops connection, server drops connection)

More formally:* All requests must start with the @ symbol. Requests not
having this will fail as a malformed packet.* Each argument must begin
with the name of the argument ("reason", "banner", etc) prefixed by a
space, and then a dash. If it has a value, it then must be given one by
placing a space and thenthe value (Values are allowed spaces).* If any
value has a - in it, the - should be escaped by replacing it with the
sequence \-(Example: String unescaped = "some -string".replace("-",
"\\-");. This is a double-backslashbecause two backslashes in a string
equals one actual backslash. 

--- Webpages ---Using this, a webpage may ban/unban a player.  **Simply
deleting a players recordfrom the database will NOT delete them (It will
only cause issues), UNLESS the serveris offline!)**.  See the above
conversation for an example of protocol.

NOTE: If you have TWO OR MORE servers running off THE SAME database, it
is IMPERATIVE that youset ALL BUT ONE of the databases to read-only mode
using the config option in config.yml! This will cause errors if you
have two servers banning to the same database, because of CACHING(The
reason MaxBans is so fast!).

--- Encryption ---Passwords are encrypted before they are sent over the
internet. Therefore, if you wishto connect using a webpage, you will
need to encrypt the password in your @connect request.The encryption
works like this (I'm trying to write it in PHP, but I haven't coded it
fora while)

function encrypt($pass){    $pass = MD5($pass); //Encrypt the password
once using MD5    $pass = $pass . "fuQJ7_q#eF78A&D"; //Append the salt
to the pass. Currently, salt is "fuQJ7_q#eF78A&D", cannot be changed.   
$pass = MD5($pass); //Encrypt the result once again    return $pass;
//This is the final result.}

--- BanManager API ---The BanManager will not request syncs with other
servers. Only the commands etc. So, if youdecide to use the BanManager
in another plugin to ban someone, the ban WILL NOT SYNC with
otherservers, unless you send the appropriate request using the
MaxBans.instance.getSyncer() object.

+++ MaxBans/src/org/maxgamer/maxbans/banmanager/BanManager.java
@@ -155,10 +155,13 @@ public class BanManager{
 			
 			//Phase 1: Load bans
 			
-			//Purge old temp bans
-			PreparedStatement ps = db.getConnection().prepareStatement("DELETE FROM bans WHERE expires <> 0 AND expires < ?");
-			ps.setLong(1, System.currentTimeMillis());
-			ps.execute(); 
+			PreparedStatement ps;
+			if(!db.isReadOnly()){
+				//Purge old temp bans
+				ps = db.getConnection().prepareStatement("DELETE FROM bans WHERE expires <> 0 AND expires < ?");
+				ps.setLong(1, System.currentTimeMillis());
+				ps.execute(); 
+			}
 			
 			plugin.getLogger().info("Loading bans");
 			query = "SELECT * FROM bans";
@@ -185,10 +188,12 @@ public class BanManager{
 			
 			//Phase 2: Load IP Bans
 			
-			//Purge old temp ip bans
-			ps = db.getConnection().prepareStatement("DELETE FROM ipbans WHERE expires <> 0 AND expires < ?");
-			ps.setLong(1, System.currentTimeMillis());
-			ps.execute(); 
+			if(!db.isReadOnly()){
+				//Purge old temp ip bans
+				ps = db.getConnection().prepareStatement("DELETE FROM ipbans WHERE expires <> 0 AND expires < ?");
+				ps.setLong(1, System.currentTimeMillis());
+				ps.execute(); 
+			}
 			
 			plugin.getLogger().info("Loading ipbans");
 			query = "SELECT * FROM ipbans";
@@ -215,10 +220,12 @@ public class BanManager{
 			
 			//Phase 3: Load Mutes
 			
-			//Purge old temp mutes
-			ps = db.getConnection().prepareStatement("DELETE FROM mutes WHERE expires <> 0 AND expires < ?");
-			ps.setLong(1, System.currentTimeMillis());
-			ps.execute(); 
+			if(!db.isReadOnly()){
+				//Purge old temp mutes
+				ps = db.getConnection().prepareStatement("DELETE FROM mutes WHERE expires <> 0 AND expires < ?");
+				ps.setLong(1, System.currentTimeMillis());
+				ps.execute();
+			}
 			
 			plugin.getLogger().info("Loading mutes");
 			query = "SELECT * FROM mutes";
@@ -278,10 +285,12 @@ public class BanManager{
 			
 			//Phase 6 loading: Load Warn history
 			
-			//Purge old warnings
-			ps = db.getConnection().prepareStatement("DELETE FROM warnings WHERE expires < ?");
-			ps.setLong(1, System.currentTimeMillis());
-			ps.execute(); 
+			if(!db.isReadOnly()){
+				//Purge old warnings
+				ps = db.getConnection().prepareStatement("DELETE FROM warnings WHERE expires < ?");
+				ps.setLong(1, System.currentTimeMillis());
+				ps.execute();
+			}
 			
 			plugin.getLogger().info("Loading warn history...");
 			//We only want warns that haven't expired.
@@ -315,7 +324,9 @@ public class BanManager{
 			//Phase 8 loading: Load history
 			plugin.getLogger().info("Loading history...");
 			
-			db.getConnection().prepareStatement("DELETE FROM history WHERE created < " + (System.currentTimeMillis() - plugin.getConfig().getInt("history-expirey-minutes", 10080) * 60000)).execute();
+			if(!db.isReadOnly()){
+				db.getConnection().prepareStatement("DELETE FROM history WHERE created < " + (System.currentTimeMillis() - plugin.getConfig().getInt("history-expirey-minutes", 10080) * 60000)).execute();
+			}
 			rs = db.getConnection().prepareStatement("SELECT * FROM history ORDER BY created DESC").executeQuery();
 			while(rs.next()){
 				this.history.add(new HistoryRecord(rs.getString("message"), rs.getLong("created")));
@@ -502,7 +513,6 @@ public class BanManager{
     	name = name.toLowerCase();
     	Ban ban = this.bans.get(name);
     	TempBan tBan = this.tempbans.get(name);
-    	String ip = this.getIP(name);
     	
     	if(ban != null){
     		this.bans.remove(name);
@@ -515,7 +525,6 @@ public class BanManager{
     			db.execute("DELETE FROM bans WHERE name = ?", name);
     		}
     	}
-    	unbanip(ip);
     }
     
     /**
@@ -701,12 +710,14 @@ public class BanManager{
     }
     
     /**
-     * Notes that a player joined from the given IP.
-     * @param name Name of player. Case insensitive.
-     * @param ip The ip they're connecting from.
+     * Notes the actual-case version of a players name in
+     * the database.
+     * @param name The name, case insensitive. This is converted to lowercase.
+     * @param actual The actual name, CASE SENSITIVE.  This is stored as is.
+     * @return True if the name has changed from last time, or there is
+     * no record of the player previously.
      */
-    public void logIP(String name, String ip){
-    	String actual = name;
+    public boolean logActual(String name, String actual){
     	name = name.toLowerCase();
     	
     	//Record the players original name versus the lowercase one.
@@ -714,13 +725,26 @@ public class BanManager{
 		if(oldActual == null){
 			//We've never seen this player before.
 			plugin.getDB().execute("INSERT INTO players (name, actual) VALUES (?, ?)", name, actual);
+			return true;
 		}
 		else if(!oldActual.equals(actual)){
 			plugin.getDB().execute("UPDATE players SET actual = ? WHERE name = ?", actual, name);
+			return true;
 		}
+		return false; //Nothing has changed.
+    }
+    
+    /**
+     * Notes that a player joined from the given IP.
+     * @param name Name of player. Case insensitive.
+     * @param ip The ip they're connecting from.
+     * @return True if their IP has changed, or they had none previously.
+     */
+    public boolean logIP(String name, String ip){
+    	name = name.toLowerCase();
     	
     	String oldIP = this.recentips.get(name);
-    	if(ip.equals(oldIP)) return; //Nothing has changed.
+    	if(ip.equals(oldIP)) return false; //Nothing has changed.
     	
     	if(!this.recentips.containsKey(name)){
     		//First time we've seen this player! Add them to the autocomplete list
@@ -750,6 +774,7 @@ public class BanManager{
     	}
     	
     	this.recentips.put(name, ip);
+    	return true;
     }
 	
 	/**
